refactor: isolate diagnostics and stow types
- move runtime redaction and report output into diagnostics
- extract shared path helpers and stow action/report types
- simplify the binary entrypoint and config construction flow
